feat(PGP_XOR): Create pgp_xor encryption and decryption
I handle those two methods but without parsing for the moment
